Iniital commit of a subduino_all_parameter, an attempt at prioritizing and sending all available SSM parameters.
A new approach. Read the ECU memory in two separate requests, with the second of request happening
less frequently as the parameters it involves are less important.
Send the SSM data over CAN packets, but prioritized and scheduled according to their importance. RPM, speed,
a few other paramters are sent most frequently, with the SSM switch parameter sent every 10 loops of the
aruduino.
Renamed the original subduino to subduino_basic.ino to preserve it, may need it again.